secured Recipe routes UPDATE,DELETE

+++ routes/recipe.routes.js
@@ -10,23 +10,24 @@ const Post = require('../models/Post.model');
 
 //CREATE new recipe
 router.post('/recipes', isAuthenticated, (req, res, next) => {
-    const { 
-        title, 
-        description, 
+    const {
+        title,
+        description,
         image,
         servings,
         ingredients,
         cookingAdvice,
         user } = req.body; //, req.body.user is just the user._id
 
-    Recipe.create({ 
-        title, 
+    Recipe.create({
+        title,
         description,
         image,
         servings,
         ingredients,
         cookingAdvice,
-        user })
+        user
+    })
         .then(response => {
             res.json(response)
         })
@@ -65,33 +66,47 @@ router.get('/recipes/:recipeId', (req, res, next) => {
 
 //UPDATE recipe -->req.body does not contain user._id (owner/author), would not be changed
 router.put('/recipes/:recipeId', isAuthenticated, (req, res, next) => {
-    if (!mongoose.Types.ObjectId.isValid(req.params.recipeId)) {
-        res.status(400).json({ message: 'Specified id is not valid' });
-        return;
-    }
-
-    Recipe.findByIdAndUpdate(req.params.recipeId, req.body, { returnDocument: 'after' })
-        .then((updatedRecipe) => res.json(updatedRecipe))
-        .catch(error => res.json(error));
-});
-
-
-
-//DELETE recipe
-router.delete('/recipes/:recipeId', isAuthenticated, (req, res, next) => {
     const { recipeId } = req.params;
-
     if (!mongoose.Types.ObjectId.isValid(recipeId)) {
         res.status(400).json({ message: 'Specified id is not valid' });
         return;
     }
 
-    Recipe.findByIdAndRemove(recipeId)
-        .then(deteletedRecipe => {
-            return Post.deleteMany({ _id: { $in: deteletedRecipe.posts } });
+    Recipe.findById(recipeId)
+        .then(recipe => {
+            if (req.payload._id == recipe.user) { //secure Backend by check if user from payload eq author of ressource
+                Recipe.findByIdAndUpdate(recipeId, req.body, { returnDocument: 'after' })
+                    .then((updatedRecipe) => res.json(updatedRecipe))
+                    .catch(error => res.json(error));
+            }
+            else res.status(401).json({ message: 'Unauthorized PUT request, you are not the Author of these ressource' });
         })
-        .then(() => res.json({ message: `Recipe with id ${recipeId} & all associated tasks were removed successfully.` }))
-        .catch(error => res.status(500).json(error));
+        .catch(error => res.json(error))
+});
+
+
+//DELETE recipe
+router.delete('/recipes/:recipeId', isAuthenticated, (req, res, next) => {
+    const { recipeId } = req.params;
+    console.log("DELETE req.payload: ", req.payload);
+    if (!mongoose.Types.ObjectId.isValid(recipeId)) {
+        res.status(400).json({ message: 'Specified id is not valid' });
+        return;
+    }
+
+    Recipe.findById(recipeId)
+        .then(recipe => {
+            if (req.payload._id == recipe.user) { //secure Backend by check if user from payload eq author of ressource
+                Recipe.findByIdAndRemove(recipeId)
+                    .then(deteletedRecipe => {
+                        return Post.deleteMany({ _id: { $in: deteletedRecipe.posts } });
+                    })
+                    .then(() => res.json({ message: `Recipe with id ${recipeId} & all associated tasks were removed successfully.` }))
+                    .catch(error => res.status(500).json(error));
+            }
+            else res.status(401).json({ message: 'Unauthorized DELETE request, you are not the Author of these ressource' });
+        })
+        .catch(error => res.json(error))
 });
 
 module.exports = router;
